refactor package names, update sonar properties
