+++ sonar.project.properties
@@ -3,4 +3,6 @@ sonar.projectKey=sawyern.cookiebot:cookiebot
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
 # This property is optional if sonar.modules is set.
-sonar.sources=.
+sonar.sources=./src
+
+sonar.coverage.exclusions=**/test/**,**/config/**,**/constants/**, **/properties/**, **/exception/**, **/models/**
