--- src/main/java/sawyern/cookiebot/constants/CookieConstants.java
@@ -1,6 +1,0 @@
-package sawyern.cookiebot.constants;
-
-public class CookieConstants {
-
-    private CookieConstants() {}
-}
