+++ src/main/java/sawyern/cookiebot/util/MapUtil.java
@@ -6,6 +6,14 @@ import java.util.*;
 
 @UtilityClass
 public class MapUtil {
+
+    /**
+     * sorts a map by value descending by converting the map to a linked hash map
+     * @param map map to sort
+     * @param <K> type of the key of the map
+     * @param <V> type of the value of the map
+     * @return LinkedHashMap sorted starting from the lowest value to highest value
+     */
     public static <K, V extends Comparable<? super V>> Map<K, V> sortByValue(Map<K, V> map) {
         List<Map.Entry<K, V>> list = new ArrayList<>(map.entrySet());
         list.sort(Map.Entry.comparingByValue(Comparator.reverseOrder()));
