+++ src/test/java/sawyern/cookiebot/commands/GenericBotCommandTest.java
@@ -1,4 +1,4 @@
-package sawyern.cookiebot.bot;
+package sawyern.cookiebot.commands;
 
 import discord4j.core.DiscordClient;
 import discord4j.core.event.EventDispatcher;
